Baseline for Fall Classic 2015
This code is the baseline for fall classic 2015.  The code includes all
manual and autonomous code developed over the past few months.  The
manual code is fully ready to use.  The autonomous code is not ready to
use.  Since the autonomous code is not ready to use the AUTONOMOUS_ON
flag (Constants.h) is set to false.  The current plan is to have
AUTONOMOUS_ON be set to false for the entire fall classic competition.